Update clean-up method in event loop and LTO setting
This commit rearranges the clean-up code in the exiting function of the event loop, and also rectifies the comment next to the Link Time Optimization (LTO) setting in Cargo.toml. Previously, fractal_app was made 'None' in Exit event, but now it's changed to 'None' along with window in exiting function to ensure a smoother shutdown. The LTO setting comment is updated to reflect its actual function in release profile.