Update fractal parameters and add Run_release_dev configuration
The fractal parameters, specifically the offset and scale, within the mandelbrot_simd.rs source file have been updated to improve the clarity and detail of the resulting images. Additionally, a new run configuration called Run_release_dev has been added to the .idea/runConfigurations directory for better testing and debugging convenience. This configuration is set to use the fractal package